source code/phi-initialization/requirements.txt

contourpy==1.2.0
cycler==0.12.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
entrypoints==0.4
frozendict==2.4.0
fsspec==2024.10.0
ftfy==6.3.1
h11==0.14.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
ipykernel==6.17.1
ipython==8.7.0
isoduration==20.11.0
jax==0.4.25
jedi==0.18.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mmcv==2.2.0
mmcv-full==1.7.2
mmengine==0.10.5
mpmath==1.3.0
multitasking==0.0.11
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.4.2
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.54.4
opencv-contrib-python==4.9.0.80
opencv-python==4.10.0.84
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.3
peewee==3.17.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
prometheus_client==0.20.0
prompt-toolkit==3.0.33
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.13.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.33.0
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.18.0
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.12.0
Send2Trash==1.8.2
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
scikit-image
sounddevice==0.4.6
soupsieve==2.5
stack-data==0.6.2
sympy==1.13.1
termcolor==2.5.0
terminado==0.18.0
threadpoolctl==3.5.0
timm==0.9.8
tinycss2==1.2.1
tokenizers==0.20.1
tomli==2.0.1
torch==2.5.1
torchvision==0.20.1
tornado==6.2
tqdm==4.66.6
traitlets==5.6.0
transformers==4.46.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.5
webcolors==1.13
xgboost

